Create test helpers for asserting nil values
- Since Go tests work by failing tests when expectations are NOT met,
these funcs help explicitly check if values are nil or non-nil, or even
if one value is nil while the other is non-nil. These three separate
cases, then, can be used to fail tests for reasons that are not as
simple as an equality check, and can often be verbose.