Add test case for when `ReadJSON` fails + add `QUERY_FILE` Env var
- This Env var allows the dev to change the pre-formatted JSON file
on the fly, BUT it also makes it more convenient to test the fail
response if the JSON file were to fail to parse correctly.
- It seems impossible to test the fail response for when `Marshal`
is unable to parse the JSON map, partly because if the JSON file reads
correctly then `Marshal` will probably always succeed. On the other
hand, `Marshal` handles any invalid JSON characters put into `Path()`
so I can't seem to cause it to error out that way either.