Add Error test helper to more simply compare types of errors
- Instead of checking for specific errors, this test helper can compare
that errors are either both nil (and therefore no error messaging) or that
a certain type of error will occur by comparing the actual error message
to a prefix string version of it, i.e. "invalid syntax" might be the
start of the error message, so check for that as a prefix.