Update `go-version` list in Github Action strategy matrix
- As of Go 1.21, updates to the `go.mod` file are non-conservative,
so it doesn't always preserve compatibility with previous version of Go.
Consequently, the `go-version` list for the Github Action should be kept
to v1.23 and v1.24, the minimum compatible version for this project based
on the `go` directive in my `go.mod`, even if the `toolchain` directive will
bundle Go v1.24.1 on build.