Drop inline styling in `navbar` component for margin `x` and `y` CSS helpers
- These margin helpers add spacing in the `x` and `y` directions which helps
not only to space the links in the `navbar` but also to move the `navbar` itself
away from the very edges of the web page and the `main` content below.
- Inline styling can be useful in some cases where programmatic DOM manipulation
is the only way to modify an element's styling (perhaps an SVG element), BUT
it's generally a difficult to maintain option that should be avoided if possible.