Add conditional handling Unix time without floating point values
- Normally expecting a Unix time number with microsecond precision, so there
should be a "." to split on; however, if there is no "." and therefore no microsecond
precision, then `strings.Split` returns a slice of length 1. Consequently,
I can let the `nanoseconds` input into `time.Unix` to default to 0, and IF the
number is split on its decimal point, THEN set the nanosecond precision.