Fix error assertions in `TestCall` and `TestNewQuery`
- Originally `test.EqualErrors` failed tests IF it returned TRUE BUT it was
never returning true because I incorrectly used `test.EqualErrors` with a pointer
like I would when using `errors.As`, letting the `nil` check AND `error` check
ALWAYS fail. HOWEVER, it should fail tests when returning FALSE because it would
indicate that the errors didn't match. Fortunately, once the additional layer of
pointers is removed, the tests show that the expected `error` matches the actual `error`
  - The `nil` check failed because a pointer to `nil` isn't `nil`
  - The `error` check failed because a pointer to an `error` pointer won't satisfy
  `errors.As`. It can't and won't unwrap or deref any underlying `error`