Update `TestNewQuery` test file names to get expected `error` typing
- `./bad.json` is properly handled by my `fileproject.JSON` and the `projectpath.File`
inside, creating a file path at the project root, e.g. `<root>/bad.json`, which
works in the case where I want to receive a `FileNotFoundError` BUT if I want
to receive a `MalformedJsonError` I need to properly link to the `bad_typing.json`
file in `util/test`. To do so, I need to consider the file path from the root folder